+++ vicon-lsl-bridge/src/gui/PreviewPanel.h
@@ -157,6 +157,7 @@ private:
     // loaded recording.
     QVector<QWidget*> profile_selection_controls_;
     QVector<QWidget*> playback_controls_;
+    QWidget* playback_area_ = nullptr;
     QComboBox* calibration_profile_combo_ = nullptr;
     QLineEdit* calibration_profile_name_edit_ = nullptr;
     QLineEdit* calibration_setup_id_edit_ = nullptr;